feat: Update LabelController.php and docker-compose.yml
The changes in LabelController.php modify the coordinates for positioning text in the PDF file generated. The y-coordinate is changed from 70 to 85.

In docker-compose.yml, the changes add port mappings for services. Port 9000 is mapped to the host for the PHP service, and port 8080 is mapped to the host for the phpmyadmin service.