feat: Add limit and offset parameters to getItemsbyUser method
Update MELISitesRepositoryInterface.php and MELISitesRepository.php to add the limit and offset parameters to the getItemsbyUser method. This allows for pagination of the results when retrieving items by user from the MercadoLibre API.

These changes enhance the functionality of the getItemsbyUser method by providing options for limiting the number of items returned and specifying the starting offset.