feat: Add 'x-format-new' header to getItemsbyUser API request
This commit adds the 'x-format-new' header to the API request made in the getItemsbyUser method of the MELISitesRepository class. This header is set to 'true' and is used to indicate that the API response should be in the new format. This change enhances the functionality of the method by ensuring that the response is in the desired format.