feat: Update Item component to pass item_id to view
Update the Item component to pass the item_id variable to the view when rendering. This ensures that the correct item_id is displayed in the item.blade.php template.

This change enhances the functionality of the Item component by correctly passing the item_id to the view, improving the accuracy of the displayed information.